refactor: replace base with gopkg base
