+++ pkg/generic/json_test/generic_test.go
@@ -21,6 +21,7 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"github.com/cloudwego/kitex/pkg/generic/thrift"
 	"math"
 	"net"
 	"reflect"
@@ -33,7 +34,6 @@ import (
 
 	"github.com/cloudwego/kitex/client/callopt"
 	"github.com/cloudwego/kitex/client/genericclient"
-	"github.com/cloudwego/kitex/internal/generic/thrift"
 	kt "github.com/cloudwego/kitex/internal/mocks/thrift"
 	"github.com/cloudwego/kitex/internal/test"
 	"github.com/cloudwego/kitex/pkg/generic"
