+++ pkg/generic/thrift/parse.go
