+++ pkg/generic/thrift/parse_test.go
